[NO-JIRA] Refactor auto publish code
