+++ scripts/release-process/util.js
@@ -16,40 +16,275 @@
  * limitations under the License.
  */
 
+/* eslint-disable no-console */
+
 const fs = require('fs');
-const path = require('path');
-const { promisify } = require('util');
+const { execSync } = require('child_process');
 
-const flatten = require('lodash/flatten');
-const globby = require('globby');
-const yaml = require('js-yaml');
+const colors = require('colors');
+const YAML = require('yaml');
 
-const CHANGELOG_GLOBS = ['packages/*/changelog.yml'];
-const GLOBBY_OPTIONS = {
-  gitignore: true,
+const {
+  RELEASE_MODE_MAJOR,
+  RELEASE_MODE_MINOR,
+  RELEASE_MODE_PATCH,
+  TITLE_MAJOR,
+  TITLE_MINOR,
+  TITLE_PATCH,
+  UNRELEASED_PATH,
+  UNRELEASED_PATH_CLEAN,
+  YAML_SECTION_MAJOR,
+  YAML_SECTION_MINOR,
+  YAML_SECTION_PATCH,
+} = require('./constants');
+
+const verbose = process.argv.includes(`--verbose`);
+const testing = process.argv.includes(`--testing`);
+
+const logVerbose = (...args) => {
+  console.log(colors.cyan(...args));
 };
 
-const readFileAsync = promisify(fs.readFile);
+const logOk = (...args) => {
+  console.log(colors.green(...args));
+};
 
-const allChangelogPaths = async () =>
-  Promise.all(CHANGELOG_GLOBS.map(glob => globby(glob, GLOBBY_OPTIONS))).then(
-    flatten,
+const logWarn = (...args) => {
+  console.log(colors.yellow(...args));
+};
+
+const logError = (...args) => {
+  console.error(colors.red(...args));
+};
+
+const arrayInsert = (arr, index, item) => arr.splice(index, 0, item);
+
+const assert = (condition, message) => {
+  if (condition) {
+    return;
+  }
+
+  logError(message);
+  process.exit(1);
+};
+
+const getLernaChanges = () => {
+  const lernaChanges = JSON.parse(
+    execSync(`npx lerna changed -l --json`).toString(),
+  );
+  assert(lernaChanges, 'Could not retrieve lerna changes.');
+  return lernaChanges.map(c => c.name);
+};
+
+const getCurrentPackageMeta = () => {
+  const packages = JSON.parse(execSync(`npx lerna ls -l --json`).toString());
+  assert(packages && packages.map, 'Failed to get current package meta');
+  return packages.map(p => ({
+    name: p.name,
+    private: p.private,
+    path: p.location,
+    currentVersion: p.version,
+  }));
+};
+
+const getDate = () => {
+  const today = new Date();
+  return today.toISOString().split('T')[0];
+};
+
+const getTitleForMode = mode => {
+  if (mode === RELEASE_MODE_MINOR) {
+    return TITLE_MINOR;
+  }
+  if (mode === RELEASE_MODE_MAJOR) {
+    return TITLE_MAJOR;
+  }
+  return TITLE_PATCH;
+};
+
+const formatDescription = description => {
+  if (!description || description.length === 0) {
+    return '';
+  }
+
+  const result = [];
+  description.forEach(d => {
+    if (d && !d.match(/^\s+$/)) {
+      result.push(`   - ${d}`);
+    }
+  });
+  return result.join('\n');
+};
+
+const updatePackageJsonFiles = changes => {
+  changes.forEach(c => {
+    const packageJsonFile = `${c.path}/package.json`;
+    const packageJsonData = JSON.parse(
+      fs.readFileSync(packageJsonFile).toString(),
+    );
+    packageJsonData.version = c.newVersion;
+    fs.writeFileSync(
+      packageJsonFile,
+      `${JSON.stringify(packageJsonData, null, '  ')}\n`,
+    );
+  });
+};
+
+const createGitTags = changes => {
+  changes.forEach(c => {
+    const tagMessage = `${c.name}@${c.newVersion}`;
+    execSync(`git tag -a ${tagMessage} -m ${tagMessage}`);
+  });
+};
+
+const resetUnreleased = () => {
+  execSync(`cp ${UNRELEASED_PATH_CLEAN} ${UNRELEASED_PATH}`);
+};
+
+const validateUnreleasedEntries = (yaml, currentPackageMeta) => {
+  if (!yaml) {
+    return false;
+  }
+  if (verbose) {
+    logVerbose(`Validating yaml`, yaml);
+  }
+  Object.keys(yaml).forEach(key => {
+    if (
+      key !== YAML_SECTION_MAJOR &&
+      key !== YAML_SECTION_MINOR &&
+      key !== YAML_SECTION_PATCH
+    ) {
+      logError(
+        `Section in ${UNRELEASED_PATH} should be one of ${YAML_SECTION_PATCH}, ${YAML_SECTION_MINOR}, or ${YAML_SECTION_MAJOR}. Found ${key}`,
+      );
+      return false;
+    }
+    yaml[key].forEach(entry => {
+      const entryName = Object.keys(entry)[0];
+      if (verbose) {
+        logVerbose(`Validating entry`, entryName);
+      }
+      const correspondingPackage = currentPackageMeta.find(
+        p => p.name === entryName,
+      );
+      if (!entryName || !correspondingPackage) {
+        logError(
+          `${UNRELEASED_PATH} contains an entry for package ${entryName} but no such package exists in the project.`,
+        );
+      }
+    });
+    return true;
+  });
+  return true;
+};
+
+const getModeFromYamlSection = yamlSection => {
+  if (yamlSection === YAML_SECTION_MINOR) {
+    return RELEASE_MODE_MINOR;
+  }
+  if (yamlSection === YAML_SECTION_MAJOR) {
+    return RELEASE_MODE_MAJOR;
+  }
+  return RELEASE_MODE_PATCH;
+};
+
+const processUnreleasedYaml = yaml => {
+  const result = [];
+  if (!yaml) {
+    return result;
+  }
+  const yamlKeys = Object.keys(yaml);
+  yamlKeys.forEach(k => {
+    const entries = yaml[k];
+    entries.forEach((e, eIndex) => {
+      const entryKeys = Object.keys(e);
+      entryKeys.forEach(ek => {
+        if (verbose) {
+          console.log(`entryKeys`, entryKeys);
+        }
+        const changes = entries[eIndex][ek];
+        const newResult = {
+          name: ek,
+          mode: getModeFromYamlSection(k),
+          description: changes,
+        };
+        result.push(newResult);
+      });
+    });
+  });
+
+  return result;
+};
+
+const parseUnreleasedFile = currentPackageMeta => {
+  const unreleasedContent = fs.readFileSync(UNRELEASED_PATH).toString();
+  assert(unreleasedContent, `Failed to read ${UNRELEASED_PATH}`);
+  const unreleasedYaml = YAML.parse(unreleasedContent);
+  if (verbose) {
+    logVerbose(`unreleasedYaml`);
+    logVerbose(YAML.stringify(unreleasedYaml));
+  }
+  assert(unreleasedYaml, `Failed to parse ${UNRELEASED_PATH}`);
+  assert(
+    validateUnreleasedEntries(unreleasedYaml, currentPackageMeta),
+    `${UNRELEASED_PATH} is not valid. Please fix it and try again.`,
   );
 
-const readChangelog = async changelogPath => {
-  const content = await readFileAsync(changelogPath, { encoding: 'utf8' });
-  return yaml.safeLoad(content);
+  return processUnreleasedYaml(unreleasedYaml);
 };
 
-const determinePackageJsonPath = changelogPath =>
-  path.join(path.dirname(changelogPath), 'package.json');
+const printOutChanges = (changes, logFn) => {
+  logOk(`\n`);
+  changes.forEach(c => {
+    logFn(
+      `${c.releaseMode} change to package ${c.name}: ${c.currentVersion} => ${
+        c.newVersion
+      }`,
+    );
+    logOk(`${formatDescription(c.description)}`);
+    logOk(``);
+  });
+};
 
-const readPackageJson = async packageJsonpath =>
-  JSON.parse(await readFileAsync(packageJsonpath));
+const printOutChangeSummary = changes => {
+  const undocumentedChanges = changes.filter(c => !c.description);
+  const documentedChanges = changes.filter(c => !!c.description);
+  if (undocumentedChanges.length > 0) {
+    logWarn(`\n`);
+    logWarn(
+      `The following changes will be published with NO CHANGELOG ENTRY! 🙀`,
+    );
+    printOutChanges(undocumentedChanges, logWarn);
+  }
+  if (documentedChanges.length > 0) {
+    console.log(``);
+    logOk(`\n`);
+    logOk(
+      `The following changes will be published and reflected in CHANGELOG.md 😎`,
+    );
+    printOutChanges(documentedChanges, logOk);
+  }
+};
 
 module.exports = {
-  allChangelogPaths,
-  determinePackageJsonPath,
-  readChangelog,
-  readPackageJson,
+  arrayInsert,
+  assert,
+  createGitTags,
+  formatDescription,
+  printOutChangeSummary,
+  getCurrentPackageMeta,
+  getDate,
+  getLernaChanges,
+  getModeFromYamlSection,
+  getTitleForMode,
+  logError,
+  logOk,
+  logVerbose,
+  logWarn,
+  parseUnreleasedFile,
+  processUnreleasedYaml,
+  resetUnreleased,
+  updatePackageJsonFiles,
+  verbose,
+  testing,
 };
